Move GroupTrait trait and InactiveMember to primitives.
