+++ pallets/contracts/src/benchmarking.rs
@@ -30,7 +30,6 @@ use wasm_instrument::parity_wasm::elements::{Instruction, ValueType};
 
 use pallet_identity::ParentDid;
 use polymesh_common_utilities::benchs::{cdd_provider, user, User, UserBuilder};
-use polymesh_common_utilities::group::GroupTrait;
 use polymesh_primitives::asset::AssetId;
 use polymesh_primitives::constants::currency::POLY;
 use polymesh_primitives::identity::limits::{
@@ -38,8 +37,8 @@ use polymesh_primitives::identity::limits::{
 };
 use polymesh_primitives::secondary_key::ExtrinsicNames;
 use polymesh_primitives::{
-    AssetPermissions, Balance, ExtrinsicName, ExtrinsicPermissions, PalletName, PalletPermissions,
-    Permissions, PortfolioId, PortfolioNumber, PortfolioPermissions,
+    traits::group::GroupTrait, AssetPermissions, Balance, ExtrinsicName, ExtrinsicPermissions,
+    PalletName, PalletPermissions, Permissions, PortfolioId, PortfolioNumber, PortfolioPermissions,
 };
 
 use crate::chain_extension::*;
