+++ pallets/identity/src/claims.rs
@@ -24,14 +24,12 @@ use frame_support::{
 use frame_system::ensure_root;
 use pallet_base::{ensure_string_limited, try_next_pre};
 
-use polymesh_common_utilities::traits::{
-    group::{GroupTrait, InactiveMember},
-    identity::{Config, RawEvent},
-};
+use polymesh_common_utilities::traits::identity::{Config, RawEvent};
 use polymesh_primitives::identity_claim::CustomClaimTypeId;
 use polymesh_primitives::{
-    protocol_fee::ProtocolOp, CddId, Claim, ClaimType, IdentityClaim, IdentityId, Scope,
-    SecondaryKey, SystematicIssuers,
+    protocol_fee::ProtocolOp,
+    traits::group::{GroupTrait, InactiveMember},
+    CddId, Claim, ClaimType, IdentityClaim, IdentityId, Scope, SecondaryKey, SystematicIssuers,
 };
 use sp_runtime::traits::{CheckedAdd, SaturatedConversion, Zero};
 use sp_std::prelude::*;
