+++ pallets/runtime/tests/src/group_test.rs
@@ -4,7 +4,7 @@ use super::{
 };
 use pallet_group::{self as group};
 use pallet_identity as identity;
-use polymesh_common_utilities::traits::group::GroupTrait;
+use polymesh_primitives::traits::group::GroupTrait;
 use polymesh_primitives::IdentityId;
 
 use frame_support::{assert_noop, assert_ok, dispatch::DispatchError};
