+++ pallets/transaction-payment/src/lib.rs
@@ -61,10 +61,11 @@ use frame_support::{
     weights::{WeightToFee, WeightToFeeCoefficient, WeightToFeePolynomial},
 };
 use frame_system::pallet_prelude::{BlockNumberFor, OriginFor};
-use polymesh_common_utilities::traits::{
-    group::GroupTrait, identity::IdentityFnTrait, relayer::SubsidiserTrait,
+use polymesh_common_utilities::traits::{identity::IdentityFnTrait, relayer::SubsidiserTrait};
+use polymesh_primitives::{
+    traits::{group::GroupTrait, CddAndFeeDetails},
+    TransactionError,
 };
-use polymesh_primitives::{traits::CddAndFeeDetails, TransactionError};
 use scale_info::TypeInfo;
 use sp_runtime::{
     traits::{
